requirements.txt
pyyaml==5.3.1 
absl-py==1.0.0
numpy==1.21.6
pandas==1.3.5 
scipy==1.7.1
matplotlib==3.5.1 
seaborn==0.11.2 
scikit-learn==1.0.2 
Pillow==9.0.1
Jinja2==3.1.0
torch==1.11.0 
torchvision==0.12.0